[*] move components in folders
