+++ src/components/Palette/Palette.css
