+++ src/components/PaletteTile/PaletteTile.css
