+++ src/components/PaletteTile/PaletteTile.js
