+++ src/containers/App.js
@@ -1,7 +1,7 @@
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 
-import App from '../components/App';
+import App from '../components/App/App';
 
 
 function mapStateToProps(state) {
