+++ src/containers/PaletteTile.js
@@ -2,7 +2,7 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 
 import { setPaletteColor } from '../actions/palette';
-import PaletteTile from '../components/PaletteTile';
+import PaletteTile from '../components/PaletteTile/PaletteTile';
 
 
 function mapStateToProps() {
